+++ lib/src/auxiliary/mp_command_option_aux.dart
@@ -255,6 +255,14 @@ class MPCommandOptionAux {
         : null;
   }
 
+  static THProjectionModeType getScrapProjectionMode(THScrap scrap) {
+    return scrap.hasOption(THCommandOptionType.projection)
+        ? (scrap.getOption(THCommandOptionType.projection)
+                  as THProjectionCommandOption)
+              .mode
+        : THProjectionModeType.plan;
+  }
+
   static bool elementTypeSupportsOptionType(
     THElement element,
     THCommandOptionType optionType,
@@ -264,54 +272,60 @@ class MPCommandOptionAux {
         : false;
   }
 
-  static final Map<String, Map<String, Map<String, Object>>> allowedSubtypes = {
-    'point': {
-      'air-draught': {
-        'default': 'undefined',
-        'subtypes': <String>{'winter', 'summer', 'undefined'},
-      },
-      'station': {
-        'default': 'temporary',
-        'subtypes': <String>{'temporary', 'painted', 'natural', 'fixed'},
-      },
-      'water-flow': {
-        'default': 'permanent',
-        'subtypes': <String>{'permanent', 'intermittent', 'paleo'},
-      },
-    },
-    'line': {
-      'border': {
-        'default': 'visible',
-        'subtypes': <String>{'invisible', 'presumed', 'temporary', 'visible'},
-      },
-      'survey': {
-        'default': 'cave',
-        'subtypes': <String>{'cave', 'surface'},
-      },
-      'wall': {
-        'default': 'bedrock',
-        'subtypes': <String>{
-          'bedrock',
-          'blocks',
-          'clay',
-          'debris',
-          'flowstone',
-          'ice',
-          'invisible',
-          'moonmilk',
-          'overlying',
-          'pebbles',
-          'pit',
-          'presumed',
-          'sand',
-          'underlying',
-          'unsurveyed',
+  static final Map<String, Map<String, Map<String, Object>>> supportedSubtypes =
+      {
+        'point': {
+          'air-draught': {
+            'default': 'undefined',
+            'subtypes': <String>{'winter', 'summer', 'undefined'},
+          },
+          'station': {
+            'default': 'temporary',
+            'subtypes': <String>{'temporary', 'painted', 'natural', 'fixed'},
+          },
+          'water-flow': {
+            'default': 'permanent',
+            'subtypes': <String>{'permanent', 'intermittent', 'paleo'},
+          },
         },
-      },
-      'water-flow': {
-        'default': 'permanent',
-        'subtypes': <String>{'permanent', 'conjectural', 'intermittent'},
-      },
-    },
-  };
+        'line': {
+          'border': {
+            'default': 'visible',
+            'subtypes': <String>{
+              'invisible',
+              'presumed',
+              'temporary',
+              'visible',
+            },
+          },
+          'survey': {
+            'default': 'cave',
+            'subtypes': <String>{'cave', 'surface'},
+          },
+          'wall': {
+            'default': 'bedrock',
+            'subtypes': <String>{
+              'bedrock',
+              'blocks',
+              'clay',
+              'debris',
+              'flowstone',
+              'ice',
+              'invisible',
+              'moonmilk',
+              'overlying',
+              'pebbles',
+              'pit',
+              'presumed',
+              'sand',
+              'underlying',
+              'unsurveyed',
+            },
+          },
+          'water-flow': {
+            'default': 'permanent',
+            'subtypes': <String>{'permanent', 'conjectural', 'intermittent'},
+          },
+        },
+      };
 }